Fix: seperate setting files
